requirements.txt
resend
flask
python-dotenv
anthropic